requirements.txt
python-telegram-bot==20.7
fastapi
uvicorn
requests
tzdata
beautifulsoup4
python-dotenv
